requirements.txt
# Core scientific computing
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# API clients
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Optional: Progress bars
tqdm>=4.66.0
